Invitation - Code and UUID
Invitation - Code and UUID
