+++ app/Data/Models/Invitation.php
@@ -2,6 +2,8 @@
 
 namespace App\Data\Models;
 
+use Ramsey\Uuid\Uuid;
+
 class Invitation extends Base
 {
     /**
@@ -29,5 +31,26 @@ class Invitation extends Base
     public function subEvent()
     {
         return $this->belongsTo(SubEvent::class);
+    }
+
+    public function save(array $options = [])
+    {
+        $this->code_invitation = $this->codeInvitationGenerator();
+        $this->uuid_invitation = (string) Uuid::uuid4();
+        return parent::save($options);
+    }
+
+    private function codeInvitationGenerator()
+    {
+        do {
+            $codeInvitation = collect(
+                array_merge(
+                    array_random(range('A', 'Z'), 3),
+                    array_random(range(0, 9), 3)
+                )
+            )->implode('');
+        } while ($this->where('code_invitation', $codeInvitation)->first());
+
+        return $codeInvitation;
     }
 }
